requirements.txt
Django==4.2.6
django-oauth-toolkit==2.3.*
djangorestframework==3.14.*
drf-spectacular==0.26.5
